+++ CHANGELOG.md
@@ -2,6 +2,37 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.4.0](https://git.hyperchain.cn///compare/v1.4.0-3...v1.4.0) (2022-04-26)
+
+## [1.4.0-3](https://git.hyperchain.cn///compare/v1.4.0-2...v1.4.0-3) (2022-04-06)
+
+
+### Bug Fixes
+
+* **transactionHash:** #QAGC-161, fix getTransaction method for hyperchain ([0870517](https://git.hyperchain.cn///commit/0870517bba5fa5960f10107b2db92ac9121cecc0))
+
+## [1.4.0-2](https://git.hyperchain.cn///compare/v1.4.0-1...v1.4.0-2) (2022-03-29)
+
+
+### Bug Fixes
+
+* **snapshot:** change flato version snapshot() underlying call as archive_makeSnapshot4Flato ([c896e46](https://git.hyperchain.cn///commit/c896e465ccff00a09d3e2f3f7a5d6e0fc39ef6a3))
+
+## [1.4.0-1](https://git.hyperchain.cn///compare/v1.3.0...v1.4.0-1) (2022-03-25)
+
+
+### Features
+
+* **hashChange:** suport hashChangeContract of bvm ([18ff8db](https://git.hyperchain.cn///commit/18ff8db2d168501b656c6d6322da370d0f8aa986))
+* **proof:** add tx, account and state proof ([f332bce](https://git.hyperchain.cn///commit/f332bceb9720fb2e9e6359a0508e398e70f7cd93))
+
+
+### Bug Fixes
+
+* **Abi:** fix Abi's toJson method ([368961d](https://git.hyperchain.cn///commit/368961d6807428aa515b3e73b9c3670097f67ac6))
+* **request:** fix NullPointEx for isUsedProvider method ([b3b6e73](https://git.hyperchain.cn///commit/b3b6e734abf061b379d7daab0281c6b86050f942))
+* **TxService:** set ns for sendBatchTxs ([966868e](https://git.hyperchain.cn///commit/966868ef312d9f46eed93c78cabecf109c1235f8))
+
 ## [1.3.0](https://git.hyperchain.cn///compare/v1.3.0-3...v1.3.0) (2022-03-25)